Restoring systemd

+++ replace_n_launch.sh
@@ -2,6 +2,10 @@
 # this should be run as root
 set -x
 
+command_exists() {
+	command -v "$@" > /dev/null 2>&1
+}
+
 case "$(uname)" in
 Darwin)
     # TODO: Darwin is not running today - fix launchd
@@ -16,6 +20,7 @@ Darwin)
 Linux)
     if [ -x /etc/init.d/webexec ]; then
         /etc/init.d/webexec stop
+        rm /etc/init.d/webexec
     fi
     if [ -f /etc/systemd/system/webexec.service ]; then
         systemctl stop webexec.service
@@ -23,9 +28,19 @@ Linux)
     fi
     cp webexec /usr/local/bin
     sh -c "echo USER=$1 >/etc/webexec"
-    cp webexecd.sh /etc/init.d/webexec
-    chown root:root /etc/init.d/webexec
-    update-rc.d webexec defaults
-    /etc/init.d/webexec start 
+    if command_exists systemctl; then
+        cp webexec.service.tmpl webexec.service
+        sed -i "s/\$USER/$1/; s/\$HOME/$2" webexec.service
+        chown root:root webexec.service
+        mv webexec.service /etc/systemd/system/webexec.service
+        systemctl daemon-reload
+        systemctl enable webexec.service
+        systemctl start webexec.service
+    else
+        cp webexecd.sh /etc/init.d/webexec
+        chown root:root /etc/init.d/webexec
+        update-rc.d webexec defaults
+        /etc/init.d/webexec start 
+    fi
     ;;
 esac
